requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
openai==1.11.0
python-dotenv==1.0.1
bcrypt==4.0.1
requests==2.31.0
textblob==0.17.1
schedule==1.2.0
reportlab==4.0.7
werkzeug==3.0.1 
